Go back to using `Seq` alias. Provide implicit conversions from the `scala.collection.Seq` data type only in Scala 2.13 and up.
