+++ almond/src/main/scala/plotly/Almond.scala
@@ -4,7 +4,6 @@ import java.lang.{Boolean => JBoolean, Double => JDouble, Integer => JInt}
 
 import almond.interpreter.api.{DisplayData, OutputHandler}
 
-import scala.collection.immutable.Seq
 import scala.util.Random
 import plotly.element._
 import plotly.layout._
