+++ core/shared/src/main/scala/plotly/Trace.scala
@@ -1,6 +1,5 @@
 package plotly
 
-import scala.collection.immutable.Seq
 import scala.language.implicitConversions
 
 import java.lang.{ Boolean => JBoolean, Double => JDouble }
