+++ core/shared/src/main/scala/plotly/element/Error.scala
@@ -1,8 +1,6 @@
 package plotly
 package element
 
-import scala.collection.immutable.Seq
-
 import dataclass.data
 
 import java.lang.{ Boolean => JBoolean, Double => JDouble }
