+++ core/shared/src/main/scala/plotly/element/OneOrSeq.scala
@@ -1,7 +1,5 @@
 package plotly.element
 
-import scala.collection.immutable.Seq
-
 sealed abstract class OneOrSeq[T] extends Product with Serializable
 
 object OneOrSeq {
