+++ core/shared/src/main/scala/plotly/layout/Layout.scala
@@ -5,7 +5,6 @@ import java.lang.{ Integer => JInt, Double => JDouble, Boolean => JBoolean }
 
 import plotly.element._
 import dataclass.data
-import scala.collection.immutable.Seq
 
 @data class Layout(
           title: Option[String],
