+++ demo/src/main/scala/plotly/demo/DemoChart.scala
@@ -3,8 +3,6 @@ package plotly.demo
 import plotly.Trace
 import plotly.layout.Layout
 
-import scala.collection.immutable.Seq
-
 trait DemoChart {
   def plotlyDocUrl: String
   def id: String
