+++ joda-time/src/main/scala/plotly/Joda.scala
@@ -1,6 +1,5 @@
 package plotly
 
-import scala.collection.immutable.Seq
 import org.joda.time._
 import plotly.Sequence.DateTimes
