+++ render/jvm/src/main/scala/plotly/Plotly.scala
@@ -13,7 +13,6 @@ import argonaut.{Json, PrettyParams}
 import plotly.internals.{BetterPrinter, Properties}
 
 import scala.annotation.tailrec
-import scala.collection.immutable.Seq
 
 object Plotly {
