+++ render/shared/src/main/scala/plotly/internals/ArgonautCodecsExtra.scala
@@ -1,7 +1,6 @@
 package plotly.internals
 
 import argonaut.{DecodeJson, EncodeJson}
-import scala.collection.immutable.Seq
 
 trait ArgonautCodecsExtra {
