+++ render/shared/src/main/scala/plotly/internals/ArgonautCodecsInternals.scala
@@ -9,7 +9,6 @@ import argonaut.ArgonautShapeless._
 import argonaut.derive._
 import shapeless._
 
-import scala.collection.immutable.Seq
 import scala.util.Try
 import plotly.element._
 import plotly.layout._
